refactor: consolidate InstructionBuilder into router_client module
Move InstructionBuilder trait and InstructionBuilderExt from
instructions.rs and syntax_helpers.rs into router_client/mod.rs.
Delete the now-empty modules. Update all imports.
