+++ hylo-clients/src/prelude.rs
@@ -7,11 +7,11 @@ pub use fix::prelude::*;
 pub use hylo_core::idl::tokens::{HYUSD, JITOSOL, SHYUSD, XSOL};
 
 pub use crate::exchange_client::ExchangeClient;
-pub use crate::instructions::InstructionBuilder;
 pub use crate::program_client::{ProgramClient, VersionedTransactionData};
-pub use crate::router_client::RouterClient;
+pub use crate::router_client::{
+  InstructionBuilder, InstructionBuilderExt, RouterClient,
+};
 pub use crate::stability_pool_client::StabilityPoolClient;
-pub use crate::syntax_helpers::InstructionBuilderExt;
 pub use crate::transaction::{
   BuildTransactionData, RouterArgs, TransactionSyntax,
 };
