--- hylo-clients/src/syntax_helpers.rs
@@ -1,64 +1,0 @@
-//! Extension traits for cleaner static dispatch syntax.
-
-use anchor_client::solana_sdk::instruction::Instruction;
-use anchor_client::solana_sdk::pubkey::Pubkey;
-use anyhow::Result;
-use hylo_idl::tokens::TokenMint;
-
-use crate::instructions::InstructionBuilder;
-
-/// Turbofish syntax for [`InstructionBuilder`].
-///
-/// ```rust,no_run
-/// use hylo_clients::prelude::*;
-/// use hylo_idl::tokens::{HYUSD, JITOSOL};
-///
-/// # fn example() -> Result<()> {
-/// let user = Pubkey::new_unique();
-/// let args = RouterArgs { amount: 1_000_000_000, user, slippage_config: None };
-/// let instructions = RouterClient::build_instructions::<JITOSOL, HYUSD>(args)?;
-/// let luts = RouterClient::lookup_tables::<JITOSOL, HYUSD>();
-/// # Ok(())
-/// # }
-/// ```
-pub trait InstructionBuilderExt {
-  /// Builds instructions for a token pair operation.
-  ///
-  /// # Errors
-  /// * Underlying builder errors.
-  fn build_instructions<IN, OUT>(
-    inputs: <Self as InstructionBuilder<IN, OUT>>::Inputs,
-  ) -> Result<Vec<Instruction>>
-  where
-    Self: InstructionBuilder<IN, OUT>,
-    IN: TokenMint,
-    OUT: TokenMint;
-
-  fn lookup_tables<IN, OUT>() -> &'static [Pubkey]
-  where
-    Self: InstructionBuilder<IN, OUT>,
-    IN: TokenMint,
-    OUT: TokenMint;
-}
-
-impl<X> InstructionBuilderExt for X {
-  fn build_instructions<IN, OUT>(
-    inputs: <Self as InstructionBuilder<IN, OUT>>::Inputs,
-  ) -> Result<Vec<Instruction>>
-  where
-    Self: InstructionBuilder<IN, OUT>,
-    IN: TokenMint,
-    OUT: TokenMint,
-  {
-    <Self as InstructionBuilder<IN, OUT>>::build(inputs)
-  }
-
-  fn lookup_tables<IN, OUT>() -> &'static [Pubkey]
-  where
-    Self: InstructionBuilder<IN, OUT>,
-    IN: TokenMint,
-    OUT: TokenMint,
-  {
-    <Self as InstructionBuilder<IN, OUT>>::REQUIRED_LOOKUP_TABLES
-  }
-}
