+++ hylo-quotes/src/protocol_state_strategy/router.rs
@@ -4,8 +4,7 @@ use anchor_lang::prelude::Pubkey;
 use anyhow::Result;
 use async_trait::async_trait;
 use fix::prelude::*;
-use hylo_clients::router_client::RouterClient;
-use hylo_clients::syntax_helpers::InstructionBuilderExt;
+use hylo_clients::router_client::{InstructionBuilderExt, RouterClient};
 use hylo_clients::transaction::RouterArgs;
 use hylo_core::slippage_config::SlippageConfig;
 use hylo_core::solana_clock::SolanaClock;
