+++ hylo-quotes/src/simulation_strategy/router.rs
@@ -5,8 +5,7 @@ use anyhow::Result;
 use async_trait::async_trait;
 use fix::prelude::{UFix64, N4, N6, N8, N9};
 use fix::typenum::Integer;
-use hylo_clients::router_client::RouterClient;
-use hylo_clients::syntax_helpers::InstructionBuilderExt;
+use hylo_clients::router_client::{InstructionBuilderExt, RouterClient};
 use hylo_clients::transaction::RouterArgs;
 use hylo_core::slippage_config::SlippageConfig;
 use hylo_core::solana_clock::SolanaClock;
